Move all literals test to test_utils
